[FIXUP] KVM: guest_memfd: Allow merging folios after to-private conversion
kvm_gmem_restructure() is invoked to merge smaller folios during
shared-to-private conversions, when the entire range of gi->page_order will
be all private after the shared-to-private conversion.

However, at the point of invoking kvm_gmem_restructure(), the range to be
converted to private is not yet set to private. Thus, checking of whether
the entire gi->page_order range is private couldn't return true. Merging of
smaller folios therefore couldn't occur.

Since setting the conversion range's attribute to private is immediately
after kvm_gmem_restructure(), (i.e., the range to be converted must be set
to private after kvm_gmem_restructure() succeeds), do not check the private
attribute of the range to be converted and instead just check if the rest
parts in the entire range of gi->page_order are all private to determine
the merge.

Signed-off-by: Yan Zhao <yan.y.zhao@intel.com>